Move 'Payments' to Provider workspace
