+++ src/Server.UI/Pages/Workspaces/Provider/Components/CumulativeDetailsComponent.razor
@@ -1,6 +1,6 @@
 ﻿@using Cfo.Cats.Application.Features.Payments.Extensions
 @using Cfo.Cats.Server.UI.Pages.Dashboard.Components
-@using Cfo.Cats.Server.UI.Pages.Payments.Components
+@using Cfo.Cats.Server.UI.Pages.Workspaces.Provider.Components.Payments
 @inherits CatsComponentBase
 
 @if (_isLoading)
