+++ src/Server.UI/Pages/Workspaces/Provider/Components/Payments/ActivityPayments.razor.cs
@@ -6,12 +6,16 @@ using Cfo.Cats.Application.Features.Payments.DTOs;
 using Cfo.Cats.Application.Features.Payments.Queries;
 using Cfo.Cats.Infrastructure.Constants;
 
-namespace Cfo.Cats.Server.UI.Pages.Payments.Components;
+namespace Cfo.Cats.Server.UI.Pages.Workspaces.Provider.Components.Payments;
 
-public partial class InductionPayments
+public partial class ActivityPayments
 {
-    private bool _loading = true;
+    private readonly ApexChartOptions<ActivityPaymentSummaryDto> _options = new();
+    private bool _loading;
     private bool _downloading;
+
+    [CascadingParameter(Name = "IsDarkMode")]
+    public bool IsDarkMode { get; set; }
 
     [Parameter, EditorRequired] public bool DataView { get; set; }
 
@@ -21,24 +25,20 @@ public partial class InductionPayments
 
     [Parameter] public ContractDto? Contract { get; set; }
 
-    [CascadingParameter]
-    public UserProfile CurrentUser { get; set; } = default!;
+    [CascadingParameter] public UserProfile CurrentUser { get; set; } = default!;
 
-    private InductionPaymentDto[] Payments { get; set; } = [];
-    private List<InductionPaymentSummaryDto> SummaryData = [];
-
-    private GetInductionPayments.Query? Query;
-
-    [CascadingParameter(Name = "IsDarkMode")]
-    public bool IsDarkMode { get; set; }
-
-    public ApexChartOptions<InductionPaymentSummaryDto> Options => new()
+    public ApexChartOptions<ActivityPaymentSummaryDto> Options => new()
     {
         Theme = new Theme
         {
             Mode = IsDarkMode ? Mode.Dark : Mode.Light
         }
     };
+
+    private ActivityPaymentDto[] Payments = [];
+    private List<ActivityPaymentSummaryDto> SummaryData = [];
+
+    private GetActivityPayments.Query? Query;
 
     private async Task OnRefresh()
     {
@@ -92,7 +92,7 @@ public partial class InductionPayments
         try
         {
             _downloading = true;
-            var result = await GetNewMediator().Send(new ExportInductionPayments.Command()
+            var result = await GetNewMediator().Send(new ExportActivityPayments.Command()
             {
                 Query = Query!
             });
@@ -115,4 +115,5 @@ public partial class InductionPayments
             _downloading = false;
         }
     }
+
 }
