+++ src/Server.UI/Pages/Workspaces/Provider/Components/Payments/EnrolmentPayments.razor.cs
@@ -6,39 +6,43 @@ using Cfo.Cats.Application.Features.Payments.DTOs;
 using Cfo.Cats.Application.Features.Payments.Queries;
 using Cfo.Cats.Infrastructure.Constants;
 
-namespace Cfo.Cats.Server.UI.Pages.Payments.Components;
+namespace Cfo.Cats.Server.UI.Pages.Workspaces.Provider.Components.Payments;
 
-public partial class ActivityPayments
+public partial class EnrolmentPayments
 {
-    private readonly ApexChartOptions<ActivityPaymentSummaryDto> _options = new();
-    private bool _loading;
+    private bool _loading = true;
     private bool _downloading;
+
+    [Parameter, EditorRequired]
+    public bool DataView { get; set; }
+
+    [Parameter, EditorRequired]
+    public int Month { get; set; }
+
+    [Parameter, EditorRequired]
+    public int Year { get; set; }
+
+    [Parameter]
+    public ContractDto? Contract { get; set; }
+
+    [CascadingParameter]
+    public UserProfile CurrentUser { get; set; } = default!;
+
+    private EnrolmentPaymentDto[] Payments { get; set; } = [];
+    private List<EnrolmentPaymentSummaryDto> SummaryData = [];
+
+    private GetEnrolmentPayments.Query? Query;
 
     [CascadingParameter(Name = "IsDarkMode")]
     public bool IsDarkMode { get; set; }
 
-    [Parameter, EditorRequired] public bool DataView { get; set; }
-
-    [Parameter, EditorRequired] public int Month { get; set; }
-
-    [Parameter, EditorRequired] public int Year { get; set; }
-
-    [Parameter] public ContractDto? Contract { get; set; }
-
-    [CascadingParameter] public UserProfile CurrentUser { get; set; } = default!;
-
-    public ApexChartOptions<ActivityPaymentSummaryDto> Options => new()
+    public ApexChartOptions<EnrolmentPaymentSummaryDto> Options => new()
     {
         Theme = new Theme
         {
             Mode = IsDarkMode ? Mode.Dark : Mode.Light
         }
     };
-
-    private ActivityPaymentDto[] Payments = [];
-    private List<ActivityPaymentSummaryDto> SummaryData = [];
-
-    private GetActivityPayments.Query? Query;
 
     private async Task OnRefresh()
     {
@@ -92,7 +96,7 @@ public partial class ActivityPayments
         try
         {
             _downloading = true;
-            var result = await GetNewMediator().Send(new ExportActivityPayments.Command()
+            var result = await GetNewMediator().Send(new ExportEnrolmentPayments.Command()
             {
                 Query = Query!
             });
