+++ src/Server.UI/Pages/Workspaces/Provider/Components/Payments/_Imports.razor
@@ -2,7 +2,7 @@
 @using Cfo.Cats.Application.Features.ManagementInformation
 @using Cfo.Cats.Domain.Common.Enums
 @using Ardalis.SmartEnum;
-@using Cfo.Cats.Server.UI.Pages.Payments.Components;
+@using Cfo.Cats.Server.UI.Pages.Workspaces.Provider.Components.Payments;
 @using ApexCharts
 
 @inject ITargetsProvider TargetProvider;
