+++ src/Server.UI/Pages/Workspaces/Provider/Components/ProviderPageHeader.razor
@@ -30,6 +30,7 @@
             "case-management" => [home, ProviderLinks.CaseManagement.AsMudBreadcrumb()],
             "performance" => [home, ProviderLinks.Performance.AsMudBreadcrumb()],
             "cumulatives" => [home, ProviderLinks.Cumulatives.AsMudBreadcrumb()],
+            "payments" => [home, ProviderLinks.Payments.AsMudBreadcrumb()],
             _ => throw new Exception()
         };
     }
