+++ src/Server.UI/Pages/Workspaces/Provider/Pages/Index.razor.cs
@@ -24,6 +24,7 @@ public partial class Index
 
         var hasAdditionalRoles = (await AuthService.AuthorizeAsync(authState.User, SecurityPolicies.UserHasAdditionalRoles)).Succeeded;
         var canViewCumulatives = authState.User.IsInRole(RoleNames.SystemSupport) || authState.User.IsInRole(RoleNames.Finance);
+        var canViewPayments = authState.User.IsInRole(RoleNames.SystemSupport) || authState.User.IsInRole(RoleNames.Finance);
 
         _showCaseWorkload = hasAdditionalRoles;
 
@@ -43,6 +44,11 @@ public partial class Index
         if (canViewCumulatives)
         {
             links.Add(ProviderLinks.Cumulatives);
+        }
+
+        if (canViewPayments)
+        {
+            links.Add(ProviderLinks.Payments);
         }
 
         Links = links.ToArray();
