+++ src/Server.UI/Pages/Workspaces/Provider/Services/ProviderLinks.cs
@@ -9,4 +9,5 @@ public static class ProviderLinks
     public static BreadcrumbLinkModel CaseManagement => new ( "Case Management", "Location, pathway plan and initiative breakdowns for your cases" , $"{Home.Href}/case-management");
     public static BreadcrumbLinkModel Performance => new ( "Performance", "Enrolment, activity and outcome performance for your cases" , $"{Home.Href}/performance");
     public static BreadcrumbLinkModel Cumulatives => new ( "Cumulatives", "Cumulative performance figures against contract targets" , $"{Home.Href}/cumulatives");
+    public static BreadcrumbLinkModel Payments => new ( "Payments", "Contract payment auditing and management" , $"{Home.Href}/payments");
 }
